housekeeping, still ironing kinks
The value of deadend pores needs to be decided.
